Fix dragstop and resizestop not fired
In chrome touch mode, `dragstop` and `resizestop` were not fired immediately after releasing an element. This is because the global documentElement touchend event was never fired when you started the touch on an element.

By adding a touchend event to the element, we make sure that the up method is called. This effectively emmits the `dragstop` and `resizestop` events immediately.